Neural Network model to predict stock price
I build a Neural Network model with more than 110 economic signals and 100 stocks to predict stock prices. Based on the prediction, I picked up predicted top 10 and bottom 10 stocks to build the best and worst portfolios. As I expected, the best portfolio outperformed the worst one, but the difference between the two portfolios was not extreme because the pool of stock selection was restricted, hyperparameters of the model were not absolutely adjusted, and the number of iterations was not enough to train the model.
I can infer that a Neural Network model requires a huge amount of computing power and an adjustment process to perform well. Furthermore, the main problem of a neural network is that it is hard to interpret the economic impact of each independent variable.