Forecasting Stock Price with VAR and LASSO
I build a stock price prediction model with VAR and LASSO methods.
First, Check Granger Causality of all possible combinations of the Time series. The rows are the response variable, columns are predictors. The values in the table are the P-Values. P-Values lesser than the significance level (0.05), implies the Null Hypothesis that the coefficients of the corresponding past values is zero, that is, the X does not cause Y can be rejected.
Based on the analysis of significance, I did backtest with the Vector Auto Regression model and the LASSO model. The performance of LASSO is better than VAR's with lower prediction errors.
I concluded LASSO model is not only more efficient but also more accurate than VAR model.