Which stock can survive from SP500 after 20 years
S&P 500 is the most famous and crucial stock index tracking the performance of 500 companies S&P picked up. If S&P picks up stock, the pick gives a positive sign to the stock market, and S&P should buy the stocks so that the stock price will increase. If a stock is delisted in the index, the exclusion presents a negative sign to the market, and S&P should sell the stocks. As a result of delisting, the stock price will decrease. Because enlisting and delisting from the S&P 500 can influence the stock price, predicting which company will be delisted from the index makes investors avoid loss. A stock price generally decreases about 15% in the long term after the stock is delisted from the index.
I had research purposes below.
a)	What characteristics affect delisting from S&P 500
With Random Forest and Neural Network models, I will investigate what characteristics of a company influence the delisting from the index.
b)	Which stocks will be delisted from the index
Among current stocks in the index, I will figure out which stocks are most exposed to the risk of delisting.