The sensitivity test for titanic classification models
The researcher conducted the initial analysis to solve the problem of classifying people who would survive when the Titanic was destroyed. The research uses methods DT(Decision Tree) and NN(Neural Network), based on the real Titanic passengers data with 1,045 passengers data with seven variables, class, sex, age, number of siblings, and number of parents passenger fare.
The accuracy of the DT model is 83.44% for the entire data set. While ‘sex’ is the most crucial variable for survival, both ‘class’ and ‘age’ are important, but the difference of importance between sex and the other variables is vital. Furthermore, the confusion matrix with this model shows that False positive is 2.5% and False-negative is 14%.
The accuracy of the NN model is 78.28% for the total data set. While ‘sex’ is the most important variable for survival too, ‘class,’ ‘age,’ ‘the number of siblings,’ and ‘fare’ are similarly important. However, the difference of importance between sex and the others is not significant. In addition, the confusion matrix with this model shows that False positive is 9.1% and False-negative is 13%.