Option Pricing Back Solving Problem
Sometimes backsolving method should be applied to option pricing because the value of the target asset is obscure when a company is in the early stage of its business and can not make any profit yet.
However, as the company makes a profit, its option value increases, and its asset value should be reflected in the DCF model of the asset. To set the asset value as expected in the previous option pricing phase, backsolving method is used to get target asset's value.